% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_footnotes.R
\name{add_footnotes}
\alias{add_footnotes}
\title{Inserts Footnotes in appropriate places in Word files}
\usage{
add_footnotes(
  docx_in,
  docx_out,
  figures_path,
  tables_path,
  standard_footnotes_yaml = NULL,
  include_object_path = FALSE,
  footnotes_fail_on_missing_metadata = TRUE,
  debug = F
)
}
\arguments{
\item{docx_in}{The file path to the input .docx file.}

\item{docx_out}{The file path to the output .docx file to save to.}

\item{figures_path}{The file path to the figures and associated metadata directory.}

\item{tables_path}{The file path to the tables and associated metadata directory.}

\item{standard_footnotes_yaml}{The file path to the standard_footnotes.yaml. Default is NULL. If NULL, a default standard_footnotes.yaml bundled with the reportifyr package is used.}

\item{include_object_path}{A boolean indicating whether to include the file path of the figure or table in the footnotes. Default is FALSE.}

\item{footnotes_fail_on_missing_metadata}{A boolean indicating whether to stop execution if the metadata .json file for a figure or table is missing. Default is TRUE.}

\item{debug}{Debug.}
}
\description{
Reads in a .docx file and returns a new version with footnotes placed at appropriate places in the document.
}
\examples{
\dontrun{

# ---------------------------------------------------------------------------
# Load all dependencies
# ---------------------------------------------------------------------------
docx_in <- file.path(here::here(), "report", "shell", "template.docx")
doc_dirs <- make_doc_dirs(docx_in = docx_in)
figures_path <- file.path(here::here(), "OUTPUTS", "figures")
tables_path <- file.path(here::here(), "OUTPUTS", "tables")
standard_footnotes_yaml <- file.path(here::here(), "report", "standard_footnotes.yaml")

# ---------------------------------------------------------------------------
# Step 1.
# Table addition running add_tables will format and insert tables into the doc.
# ---------------------------------------------------------------------------
add_tables(
  docx_in = docx_in,
  docx_out = doc_dirs$doc_tables,
  tables_path = tables_path
)

# ---------------------------------------------------------------------------
# Step 2.
# Next we place in the plots using the add_plots function.
# ---------------------------------------------------------------------------
add_plots(
  docx_in = doc_dirs$doc_tables,
  docx_out = doc_dirs$doc_tabs_figs,
  figures_path = figures_path
)

# ---------------------------------------------------------------------------
# Step 3.
# Now we can add the footnotes to all the inserted figures and tables.
# ---------------------------------------------------------------------------
add_footnotes(
  docx_in = doc_dirs$doc_tabs_figs,
  docx_out = doc_dirs$doc_draft,
  figures_path = figures_path,
  tables_path = tables_path,
  standard_footnotes_yaml = standard_footnotes_yaml,
  include_object_path = TRUE
  footnotes_fail_on_missing_metadata
)
}
}
