% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots.R
\name{add_plots}
\alias{add_plots}
\title{Inserts Figures in appropriate places in a Microsoft Word file}
\usage{
add_plots(
  docx_in,
  docx_out,
  figures_path,
  fig_width = NULL,
  fig_height = NULL,
  debug = FALSE
)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{docx_out}{The file path to the output \code{.docx} file to save to.}

\item{figures_path}{The file path to the figures directory.}

\item{fig_width}{A global controller. The figure width in inches. Default is \code{NULL}. If \code{NULL}, the width is determined by the figure's pixel dimensions.}

\item{fig_height}{A global controller. The figure height in inches. Default is \code{NULL}. If \code{NULL}, the height is determined by the figure's pixel dimensions.}

\item{debug}{Debug.}
}
\description{
Reads in a \code{.docx} file and returns a new version with figures placed at appropriate places in the document.
}
\examples{
\dontrun{

# ---------------------------------------------------------------------------
# Load all dependencies
# ---------------------------------------------------------------------------
docx_in <- here::here("report", "shell", "template.docx")
doc_dirs <- make_doc_dirs(docx_in = docx_in)
figures_path <- here::here("OUTPUTS", "figures")
tables_path <- here::here("OUTPUTS", "tables")
standard_footnotes_yaml <- here::here("report", "standard_footnotes.yaml")

# ---------------------------------------------------------------------------
# Step 1.
# `add_tables()` will format and insert tables into the `.docx` file.
# ---------------------------------------------------------------------------
add_tables(
  docx_in = doc_dirs$doc_in,
  docx_out = doc_dirs$doc_tables,
  tables_path = tables_path
)

# ---------------------------------------------------------------------------
# Step 2.
# Next we insert the plots using the `add_plots()` function.
# ---------------------------------------------------------------------------
add_plots(
  docx_in = doc_dirs$doc_tables,
  docx_out = doc_dirs$doc_tabs_figs,
  figures_path = figures_path
)
}
}
