% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tables_alt_text.R
\name{add_tables_alt_text}
\alias{add_tables_alt_text}
\title{Inserts alt text for tables within a Microsoft Word file.}
\usage{
add_tables_alt_text(docx_in, docx_out, tables_path = NULL, debug = FALSE)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{docx_out}{The file path to the output \code{.docx} file to save to.}

\item{tables_path}{The file path to the tables directory. Used to embed artifact hashes in alt text. Default is \code{NULL}.}

\item{debug}{Debug.}
}
\description{
Inserts alt text for tables within a Microsoft Word file.
}
\examples{
\dontrun{
add_tables_alt_text("document-tabs.docx", "document-tabs_at.docx")
}
}
