% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reviewer_guide.R
\name{build_reviewer_guide}
\alias{build_reviewer_guide}
\title{Build a reviewer's guide of file names and source paths}
\usage{
build_reviewer_guide(docx_in, docx_out = NULL, figures_path, tables_path)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{docx_out}{The file path to the output \code{.docx} file to save to. Default is \code{NULL}.}

\item{figures_path}{The file path to the figures and associated metadata directory.}

\item{tables_path}{The file path to the tables and associated metadata directory.}
}
\description{
Reads in a \code{.docx} file and returns a \code{.docx} table of file names and source paths.
}
\examples{
\dontrun{
reviewer_guide(
  docx_in = docx_in,
  docx_out = docx_out,
  figures_path = figures_path,
  tables_path = tables_path
)
}
}
