% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_python.R
\name{get_args}
\alias{get_args}
\title{Gets args for us_setup.sh}
\usage{
get_args(uv_path)
}
\arguments{
\item{uv_path}{path to uv}
}
\value{
list of args
}
\description{
Gets args for us_setup.sh
}
\keyword{internal}
