% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reviewer_guide.R
\name{get_input_datasets}
\alias{get_input_datasets}
\title{Extract dataset file names from an R script}
\usage{
get_input_datasets(script_path)
}
\arguments{
\item{script_path}{Path to the R script}
}
\value{
A character vector of input dataset file names
}
\description{
Extract dataset file names from an R script
}
