% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_type.R
\name{get_meta_type}
\alias{get_meta_type}
\title{Get meta types from standard_footnotes.yaml}
\usage{
get_meta_type(path_to_footnotes_yaml)
}
\arguments{
\item{path_to_footnotes_yaml}{The file path to the \code{standard_footnotes.yaml} file.}
}
\value{
A list of \code{meta_type} to be called while performing an analysis
}
\description{
Get meta types from standard_footnotes.yaml
}
\examples{
\dontrun{
standard_footnotes_yaml <- here::here("report", "standard_footnotes.yaml")
meta_type <- get_meta_type(path_to_footnotes_yaml = standard_footnotes_yaml)
}
}
