% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_report_project.R
\name{initialize_report_project}
\alias{initialize_report_project}
\title{Create report directories within a project}
\usage{
initialize_report_project(
  project_dir,
  report_dir_name = NULL,
  outputs_dir_name = NULL
)
}
\arguments{
\item{project_dir}{The file path to the main project directory
where the directory structure will be created.
The directory must already exist; otherwise, an error will be thrown.}

\item{report_dir_name}{The directory name for where reports will be saved.
Default is \code{NULL}. If \code{NULL}, \code{report} will be used.}

\item{outputs_dir_name}{The directory name for where artifacts will be saved.
Default is \code{NULL}. If \code{NULL}, \code{OUTPUTS} will be used.}
}
\description{
Create report directories within a project
}
\examples{
\dontrun{
initialize_report_project(project_dir = tempdir())
}
}
