% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_metadata_files.R
\name{preview_metadata}
\alias{preview_metadata}
\title{Previews a single metadata file for an object}
\usage{
preview_metadata(file_name)
}
\arguments{
\item{file_name}{The file path of the file whose metadata you want to preview.}
}
\value{
A single row data frame consisting of metadata type and footnotes for the object supplied
}
\description{
Previews a single metadata file for an object
}
\examples{
\dontrun{
figures_path <- here::here("OUTPUTS", "figures")
plot_file_name <- "myplot.png"
preview_metadata(file.path(figures_path, plot_file_name))
}
}
