% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_metadata_files.R
\name{preview_metadata_files}
\alias{preview_metadata_files}
\title{Preview all metadata .json files in a directory}
\usage{
preview_metadata_files(file_dir)
}
\arguments{
\item{file_dir}{The file path to a directory containing metadata .json files.}
}
\value{
A data frame of metadata footnotes and meta type
}
\description{
Preview all metadata .json files in a directory
}
\examples{
\dontrun{
figures_path <- here::here("OUTPUTS", "figures")
preview_metadata_file(figures_path)
}
}
