% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpfy_context.R
\name{print.rpfy_context}
\alias{print.rpfy_context}
\title{Print a reportifyr context}
\usage{
\method{print}{rpfy_context}(x, ...)
}
\arguments{
\item{x}{An \code{rpfy_context} object.}

\item{...}{Unused.}
}
\value{
The context invisibly.
}
\description{
Print a reportifyr context
}
