% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_magic_strings.R
\name{remove_magic_strings}
\alias{remove_magic_strings}
\title{Removes Magic Strings from a Word file}
\usage{
remove_magic_strings(docx_in, docx_out)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{docx_out}{The file path to the output \code{.docx} file to save to.}
}
\description{
Reads in a .\code{docx} file and returns a new version with magic strings removed from the document.
}
\examples{
\dontrun{

# ---------------------------------------------------------------------------
# Load all dependencies
# ---------------------------------------------------------------------------
docx_in <- here::here("report", "shell", "template.docx")
doc_dirs <- make_doc_dirs(docx_in = docx_in)
figures_path <- here::here("OUTPUTS", "figures")
tables_path <- here::here("OUTPUTS", "tables")
standard_footnotes_yaml <- here::here("report", "standard_footnotes.yaml")

# ---------------------------------------------------------------------------
# Step 1.
# Run the `build_report()` wrapper function to replace figures, tables, and
# footnotes in a `.docx` file.
# ---------------------------------------------------------------------------
build_report(
  docx_in = doc_dirs$doc_in,
  docx_out = doc_dirs$doc_draft,
  figures_path = figures_path,
  tables_path = tables_path,
  standard_footnotes_yaml = standard_footnotes_yaml
)

# ---------------------------------------------------------------------------
# Step 2.
# Clean the output for final document creation. This will remove the ties
# between reportifyr and the document, so please be mindful!
# ---------------------------------------------------------------------------
remove_magic_strings(
  docx_in = doc_dirs$doc_draft,
  docx_out = doc_dirs$doc_final
)
}
}
\keyword{internal}
