% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_two_numbers.R
\name{sum_two_numbers}
\alias{sum_two_numbers}
\title{Sum two numbers}
\usage{
sum_two_numbers(a, b)
}
\arguments{
\item{a}{First number.}

\item{b}{Second number.}
}
\value{
A numeric value equal to \code{a + b}.
}
\description{
Returns the sum of two numeric values.
}
\examples{
sum_two_numbers(2, 3)
}
