% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reviewer_guide.R
\name{update_reviewer_guide_with_modelling}
\alias{update_reviewer_guide_with_modelling}
\title{Update reviewer guide document with modelling section}
\usage{
update_reviewer_guide_with_modelling(
  docx_in,
  docx_out,
  modelling_table,
  section_title
)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{docx_out}{The file path to the output \code{.docx} file.}

\item{modelling_table}{A formatted flextable object.}

\item{section_title}{The title for the modelling section.}
}
\value{
Invisibly returns the modelling data frame.
}
\description{
Adds the modelling flextable to the end of an existing reviewer guide document.
}
\keyword{internal}
