% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_alt_text_magic_strings.R
\name{validate_alt_text_magic_strings}
\alias{validate_alt_text_magic_strings}
\title{validate_alt_text_magic_strings}
\usage{
validate_alt_text_magic_strings(docx_in, config_yaml = NULL, debug = FALSE)
}
\arguments{
\item{docx_in}{path to docx in file}

\item{config_yaml}{path to config.yaml file}

\item{debug}{boolean, whether to open browser}
}
\description{
validate_alt_text_magic_strings
}
\examples{
\dontrun{
validate_alt_text_magic_strings("template.docx")
}
}
