% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config}
\alias{validate_config}
\title{Validate config.yaml file}
\usage{
validate_config(path_to_config_yaml)
}
\arguments{
\item{path_to_config_yaml}{The file path to the \code{config.yaml}.}
}
\value{
bool \code{TRUE} if config is valid, \code{FALSE} otherwise.
}
\description{
Validate config.yaml file
}
\examples{
\dontrun{
validate_config(here::here("report/config.yaml"))
}
}
