% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_docx.R
\name{validate_docx}
\alias{validate_docx}
\title{Validates input docx to ensure proper functionality with reportiryf}
\usage{
validate_docx(docx_in, config_yaml)
}
\arguments{
\item{docx_in}{path to input docx}

\item{config_yaml}{path to config.yaml file}
}
\description{
Validates input docx to ensure proper functionality with reportiryf
}
\examples{
\dontrun{
validate_docx(
  here::here("report/shell/template.docx"),
  here::here("report/config.yaml")
)
}
}
