% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input_args.R
\name{validate_input_args}
\alias{validate_input_args}
\title{Validates input and output arguments for reportifyr functions}
\usage{
validate_input_args(docx_in, docx_out)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{docx_out}{The file path to the output \code{.docx} file to save to.}
}
\value{
()
}
\description{
Validates input and output arguments for reportifyr functions
}
\examples{
\dontrun{
validate_input_args(
  "template.docx",
  "template-figs.docx"
)
}
}
