% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_meta.R
\name{validate_script_source}
\alias{validate_script_source}
\title{Validate a \code{script_source} object}
\usage{
validate_script_source(x)
}
\arguments{
\item{x}{A \code{script_source} object.}
}
\value{
\code{x} invisibly.
}
\description{
Errors if \code{x} is not a \code{script_source} or if any of
\code{path}, \code{creation_author}, \code{latest_author}, \code{creation_time},
\code{latest_time} are not non-empty character scalars.
}
