% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plots_alt_text.R
\name{add_plots_alt_text}
\alias{add_plots_alt_text}
\title{add_plots_alt_text()}
\usage{
add_plots_alt_text(docx_in, docx_out, debug = FALSE)
}
\arguments{
\item{docx_in}{path to document}

\item{docx_out}{path to output document with figures alt text}

\item{debug}{boolean turns on browser}
}
\description{
add_plots_alt_text()
}
\examples{
\dontrun{
add_plots_alt_text("doc-figs.docx", "doc-draft.docx")
}
}
