% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tables.R
\name{add_tables}
\alias{add_tables}
\title{Adds tables by looking for magic string}
\usage{
add_tables(docx_in, docx_out, tables_path, debug = F)
}
\arguments{
\item{docx_in}{The file path to the input .docx file.}

\item{docx_out}{The file path to the output .docx file to save to.}

\item{tables_path}{The file path to the tables and associated metadata directory.}

\item{debug}{Debug.}
}
\description{
Adds tables by looking for magic string
}
\examples{
\dontrun{
# ---------------------------------------------------------------------------
# Load all dependencies
# ---------------------------------------------------------------------------
docx_in <- file.path(here::here(), "report", "shell", "template.docx")
doc_dirs <- make_doc_dirs(docx_in = docx_in)
figures_path <- file.path(here::here(), "OUTPUTS", "figures")
tables_path <- file.path(here::here(), "OUTPUTS", "tables")
footnotes <- file.path(here::here(), "report", "standard_footnotes.yaml")

# ---------------------------------------------------------------------------
# Step 1.
# Table addition running add_tables will format and insert tables into the doc.
# ---------------------------------------------------------------------------
add_tables(
  docx_in = doc_dirs$doc_clean,
  docx_out = doc_dirs$doc_tables,
  tables_path = tables_path
)
}
}
