% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tables_alt_text.R
\name{add_tables_alt_text}
\alias{add_tables_alt_text}
\title{add_tables_alt_text}
\usage{
add_tables_alt_text(docx_in, docx_out, debug = FALSE)
}
\arguments{
\item{docx_in}{path to input document}

\item{docx_out}{path to output document}

\item{debug}{boolean for turning on debugging}
}
\description{
add_tables_alt_text
}
\examples{
\dontrun{
add_tables_alt_text("document-tabs.docx", "document-tabs_at.docx")
}
}
