% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_report.R
\name{build_report}
\alias{build_report}
\title{Updates a Word file to include formatted plots, tables, and footnotes}
\usage{
build_report(
  docx_in,
  docx_out = NULL,
  figures_path,
  tables_path,
  standard_footnotes_yaml = NULL,
  add_footnotes = TRUE,
  include_object_path = FALSE,
  footnotes_fail_on_missing_metadata = TRUE
)
}
\arguments{
\item{docx_in}{The file path to the input .docx file.}

\item{docx_out}{The file path to the output .docx file to save to. Default is NULL.}

\item{figures_path}{The file path to the figures and associated metadata directory.}

\item{tables_path}{The file path to the tables and associated metadata directory.}

\item{standard_footnotes_yaml}{The file path to the standard_footnotes.yaml. Default is NULL. If NULL, a default standard_footnotes.yaml bundled with the reportifyr package is used.}

\item{add_footnotes}{A boolean indicating whether to insert footnotes into the docx_in or not. Default is TRUE.}

\item{include_object_path}{A boolean indicating whether to include the file path of the figure or table in the footnotes. Default is FALSE.}

\item{footnotes_fail_on_missing_metadata}{A boolean indicating whether to stop execution if the metadata .json file for a figure or table is missing. Default is TRUE.}
}
\description{
Reads in a .docx file and returns an updated version with plots, tables, and footnotes replaced.
}
\examples{
\dontrun{

# ---------------------------------------------------------------------------
# Load all dependencies
# ---------------------------------------------------------------------------
docx_in <- file.path(here::here(), "report", "shell", "template.docx")
figures_path <- file.path(here::here(), "OUTPUTS", "figures")
tables_path <- file.path(here::here(), "OUTPUTS", "tables")
footnotes <- file.path(here::here(), "report", "standard_footnotes.yaml")

# ---------------------------------------------------------------------------
# Step 1.
# Run the wrapper function to replace figures, tables, and footnotes in a
# .docx file.
# ---------------------------------------------------------------------------
build_report(
  docx_in = docx_in,
  docx_out = doc_dirs$doc_draft,
  figures_path = figures_path,
  tables_path = tables_path,
  standard_footnotes_yaml = footnotes
)
}
}
