% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_document.R
\name{finalize_document}
\alias{finalize_document}
\title{Finalizes the document by removing magic strings and bookmarks}
\usage{
finalize_document(docx_in, docx_out = NULL)
}
\arguments{
\item{docx_in}{The file path to the input .docx file.}

\item{docx_out}{The file path to the output .docx file to save to. Default is NULL. If NULL, docx_out is assigned doc_dirs$doc_final using make_doc_dirs(docx_in = docx_in).}
}
\description{
Reads in a .docx file and returns a finalized version with magic strings and bookmarks removed.
}
\examples{
\dontrun{

# ---------------------------------------------------------------------------
# Load all dependencies
# ---------------------------------------------------------------------------
docx_in <- file.path(here::here(), "report", "shell", "template.docx")
figures_path <- file.path(here::here(), "OUTPUTS", "figures")
tables_path <- file.path(here::here(), "OUTPUTS", "tables")
footnotes <- file.path(here::here(), "report", "standard_footnotes.yaml")

# ---------------------------------------------------------------------------
# Step 1.
# Run the wrapper function build_report() to replace figures, tables, and
# footnotes in a .docx file.
# ---------------------------------------------------------------------------
build_report(
  docx_in = docx_in,
  docx_out = doc_dirs$doc_draft,
  figures_path = figures_path,
  tables_path = tables_path,
  standard_footnotes_yaml = footnotes
)

# ---------------------------------------------------------------------------
# Step 2.
# If you are happy with the report and are ready to finalize the document.
# ---------------------------------------------------------------------------
finalize_document(
  docx_in = doc_dirs$doc_draft,
  docx_out = doc_dirs$doc_final
)
}
}
