% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_venv_uv_paths.R
\name{get_venv_uv_paths}
\alias{get_venv_uv_paths}
\title{get_venv_uv_paths}
\usage{
get_venv_uv_paths()
}
\value{
list of paths to uv and venv directory for calling py scripts
}
\description{
get_venv_uv_paths
}
\examples{
\dontrun{
get_venv_uv_paths()
}
}
