% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_python.R
\name{initialize_python}
\alias{initialize_python}
\title{Initializes python virtual environment}
\usage{
initialize_python(continue = NULL)
}
\arguments{
\item{continue}{Optional argument to bypass asking user for
confirmation to install python deps}
}
\value{
invisibly the metadata_file path
}
\description{
Initializes python virtual environment
}
\examples{
\dontrun{
initialize_python()
}
}
