% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_report_project.R
\name{initialize_report_project}
\alias{initialize_report_project}
\title{Create report directories within a project}
\usage{
initialize_report_project(project_dir)
}
\arguments{
\item{project_dir}{The file path to the main project directory where the directory structure will be created. The directory must already exist; otherwise, an error will be thrown.}
}
\description{
Create report directories within a project
}
\examples{
\dontrun{
initialize_report_project(project_dir = tempdir())
}
}
