% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tables_figures_footnotes.R
\name{remove_tables_figures_footnotes}
\alias{remove_tables_figures_footnotes}
\title{Removes Tables, Figures, and Footnotes from a Word file}
\usage{
remove_tables_figures_footnotes(docx_in, docx_out)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{docx_out}{The file path to the output \code{.docx} file to save to.}
}
\description{
Reads in a \code{.docx} file and returns a new version with tables, figures, and footnotes removed from the document.
}
\examples{
\dontrun{

# ---------------------------------------------------------------------------
# Load all dependencies
# ---------------------------------------------------------------------------
docx_in <- here::here("report", "shell", "template.docx")
doc_dirs <- make_doc_dirs(docx_in = docx_in)

# ---------------------------------------------------------------------------
# Removing tables, figures, and footnotes
# ---------------------------------------------------------------------------
remove_tables_figures_footnotes(
  docx_in = doc_dirs$doc_in,
  docx_out = doc_dirs$doc_clean
)
}
}
