% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rds_with_metadata.R
\name{save_rds_with_metadata}
\alias{save_rds_with_metadata}
\title{Wrapper around the saveRDS function. Saves an object as .RDS and .RTF and captures analysis relevant metadata in a .json file}
\usage{
save_rds_with_metadata(
  object,
  file = "",
  meta_type = "NA",
  meta_equations = NULL,
  meta_notes = NULL,
  meta_abbrevs = NULL,
  table1_format = F,
  ...
)
}
\arguments{
\item{object}{The R object to serialize.}

\item{file}{The connection or name of the file where the R object is saved.}

\item{meta_type}{A string to specify the type of object. Default is "NA".}

\item{meta_equations}{A string or vector of strings representing equations to include in the metadata. Default is NULL.}

\item{meta_notes}{A string or vector of strings representing notes to include in the metadata. Default is NULL.}

\item{meta_abbrevs}{A string or vector of strings representing abbreviations to include in the metadata. Default is NULL.}

\item{table1_format}{A boolean indicating whether to apply table1-style formatting. Defaults to FALSE.}

\item{...}{Additional arguments passed to the base saveRDS() function.}
}
\description{
Extension to the saveRDS function that allows capturing object metadata as a separate .json file.
}
\examples{
\dontrun{
# Path to the analysis tables (.RDS) and metadata (.json files)
tables.path <- file.path(tempdir(), "tables")

# ---------------------------------------------------------------------------
# Save a simple table
# ---------------------------------------------------------------------------

out_name <- "01-12345-pk-theoph.csv"
save_rds_with_metadata(
  object = Theoph,
  file = file.path(tables.path, out_name)
)
}
}
