% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_report_project.R
\name{sync_report_project}
\alias{sync_report_project}
\title{Synchronizes report project with config and python
dependencies set through options. Uses .report_dir_name_init.json
to track differences.}
\usage{
sync_report_project(project_dir, report_dir_name = NULL)
}
\arguments{
\item{project_dir}{The file path to the main project directory
where the directory structure will be created.
The directory must already exist; otherwise, an error will be thrown.}

\item{report_dir_name}{The directory name for where reports will be saved,
default NULL will use "report"}
}
\description{
Synchronizes report project with config and python
dependencies set through options. Uses .report_dir_name_init.json
to track differences.
}
\examples{
\dontrun{
sync_report_project(here::here())
}
}
