% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_logger.R
\name{toggle_logger}
\alias{toggle_logger}
\title{Updates the logging level for console output. Default is set to WARN.
File logging is always at DEBUG level when log_file is provided.}
\usage{
toggle_logger(quiet = FALSE, log_file = NULL, lazy_file = FALSE)
}
\arguments{
\item{quiet}{suppresses messaging about log level.}

\item{log_file}{path to log file. If NULL, only console logging is enabled.}

\item{lazy_file}{if TRUE, delay log file creation until first log write.}
}
\description{
Updates the logging level for console output. Default is set to WARN.
File logging is always at DEBUG level when log_file is provided.
}
\examples{
\dontrun{
Sys.setenv("RPFY_VERBOSE" = "DEBUG")
toggle_logger()
}
}
