% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_alt_text_magic_strings.R
\name{validate_alt_text_magic_strings}
\alias{validate_alt_text_magic_strings}
\title{Validate alt text of figures/tables against their magic strings in a Microsoft Word file}
\usage{
validate_alt_text_magic_strings(docx_in, debug = FALSE)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{debug}{Debug.}
}
\description{
Validate alt text of figures/tables against their magic strings in a Microsoft Word file
}
\examples{
\dontrun{
validate_alt_text_magic_strings("template.docx")
}
}
