% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config}
\alias{validate_config}
\title{Validate config file}
\usage{
validate_config(path_to_config_yaml)
}
\arguments{
\item{path_to_config_yaml}{path to config.yaml file}
}
\value{
bool TRUE if config is valid, FALSE otherwise
}
\description{
Validate config file
}
\examples{
\dontrun{
validate_config(here::here("report/config.yaml"))
}
}
