% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_docx.R
\name{validate_docx}
\alias{validate_docx}
\title{Validates input Microsoft Word file to ensure proper functionality with reportifyr}
\usage{
validate_docx(docx_in, config_yaml)
}
\arguments{
\item{docx_in}{The file path to the input \code{.docx} file.}

\item{config_yaml}{The file path to the \code{config.yaml}.}
}
\description{
Validates input Microsoft Word file to ensure proper functionality with reportifyr
}
\examples{
\dontrun{
validate_docx(
  here::here("report/shell/template.docx"),
  here::here("report/config.yaml")
)
}
}
