% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input_args.R
\name{validate_input_args}
\alias{validate_input_args}
\title{validate_input_args}
\usage{
validate_input_args(docx_in, docx_out, config_yaml = NULL)
}
\arguments{
\item{docx_in}{input docx to work on}

\item{docx_out}{output docx to save to}

\item{config_yaml}{optional path to config yaml}
}
\value{
()
}
\description{
validate_input_args
}
\examples{
\dontrun{
validate_input_args(
  "template.docx",
  "template-figs.docx",
  "path/to/config.yaml"
)
}
}
