% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_object_metadata.R
\name{write_object_metadata}
\alias{write_object_metadata}
\title{Writes an object's metadata .json file}
\usage{
write_object_metadata(
  object_file,
  meta_type = NULL,
  meta_equations = NULL,
  meta_notes = NULL,
  meta_abbrevs = NULL,
  table1_format = F
)
}
\arguments{
\item{object_file}{The file path of the object to write metadata for.}

\item{meta_type}{A string to specify the type of object. Default is "NA".}

\item{meta_equations}{A string or vector of strings representing equations to include in the metadata. Default is NULL.}

\item{meta_notes}{A string or vector of strings representing notes to include in the metadata. Default is NULL.}

\item{meta_abbrevs}{A string or vector of strings representing abbreviations to include in the metadata. Default is NULL.}

\item{table1_format}{A boolean indicating whether table1 formatting is used for add_tables(). Defaults to FALSE.}
}
\description{
Writes an object's metadata .json file
}
\examples{
\dontrun{
ft <- flextable(iris)

write_object_metadata(ft, "table", file_path)
}
}
